Make context-size feature additive
Similar to std in a previous commit, the feature to select context size
should be additive.  This means if any crates wants the large context,
rmodbus should be built for it.  Only if no one wants the large context
should we opt for the smaller one.

Do this by dropping the `smallcontext` feature and instead adding a
`fullcontext` feature which is selected by default.