Add errors and helper methods
Adds two new errors to return when calling write functions on a frame
describing a read and vice versa. (This could be made impossible by
using typestate pattern but would require a more involved refactor).
Add helpers to check if an error represents a canonical Modbus error
type and to get the error number from it. Add function to ModbusFrame
to make setting the error state easier.